Fixing single item checkbox bug.
I was using the CGI::FormBuilder::Source::File part to create my
forms, however when I had a one value checkbox entry I was getting
incorrect behavior. For example with this type of entry:

cb_input:
 type: checkbox
 name: Option
 options: active=Have this item active

I would get two checkboxes instead of one. one called 'active' and the
other 'Have this item active'. I tracked this down to what looks like
a problem with the C::F::S::File.pm file that splits the options
apart. here is the patch that I did to solve the issue. Let me know if
there is a better way to do this or if this patch is unnecessary.